Added functionalities to calculate user CO2 emission and performance in a given travel
