--- src/components/buttonFunctions.js
@@ -1,16 +1,0 @@
-import TransportSelect from './TransportSelect' 
-import DestinationEntry from './DestinationEntry'
-activities = []
-const enterActivity = (event) => {
-    let activity = {
-        transport: document.getElementById('selected').value,
-        from: document.getElementById('From').value,
-        to: document.getElementById('To').value,
-    }
-    if (activity.transport !== null && activity.from !== null && activity.to !== null) {
-        return activity
-    }
-}
-document.addEventListener('DOMContentLoaded', () => {
-    document.getElementById('ActivityButton').addEventListener('click', enterActivity);
-})
